Set up shared rollup config for lit-html and lit-element
