+++ package.json
@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "scripts": {
     "bootstrap": "lerna bootstrap",
     "build": "lerna run build",
@@ -7,5 +8,13 @@
   },
   "dependencies": {
     "lerna": "^3.22.1"
+  },
+  "devDependencies": {
+    "@rollup/plugin-replace": "^2.3.3",
+    "@rollup/plugin-virtual": "^2.0.3",
+    "rollup-plugin-copy": "^3.3.0",
+    "rollup-plugin-sourcemaps": "^0.6.2",
+    "rollup-plugin-summary": "^1.2.3",
+    "rollup-plugin-terser": "^7.0.2"
   }
 }
