+++ packages/lit-element/package.json
@@ -51,7 +51,6 @@
     "@babel/plugin-transform-typescript": "^7.2.0",
     "@esm-bundle/chai": "^4.1.5",
     "@rollup/plugin-alias": "^3.1.1",
-    "@rollup/plugin-replace": "^2.3.3",
     "@types/chai": "^4.0.1",
     "@types/mocha": "^8.0.3",
     "@webcomponents/shadycss": "^1.8.0",
@@ -61,12 +60,7 @@
     "mocha": "^8.1.1",
     "prettier": "^2.0.5",
     "rollup": "^2.26.4",
-    "rollup-plugin-copy": "^3.3.0",
     "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-sourcemaps": "^0.6.2",
-    "rollup-plugin-summary": "^1.2.3",
-    "rollup-plugin-terser": "^7.0.0",
-    "terser": "^5.2.1",
     "tslint": "^6.1.3",
     "typescript": "^4.0.2"
   },
