+++ packages/lit-html/package.json
@@ -45,19 +45,12 @@
   "dependencies": {},
   "devDependencies": {
     "@esm-bundle/chai": "^4.1.5",
-    "@rollup/plugin-replace": "^2.3.3",
-    "@rollup/plugin-virtual": "^2.0.3",
     "@types/mocha": "^8.0.3",
     "chokidar-cli": "^2.1.0",
     "concurrently": "^5.3.0",
     "mocha": "^8.1.3",
     "prettier": "^2.1.1",
-    "rollup": "^2.26.9",
-    "rollup-plugin-copy": "^3.3.0",
-    "rollup-plugin-sourcemaps": "^0.6.2",
-    "rollup-plugin-summary": "^1.2.3",
-    "rollup-plugin-terser": "^7.0.1",
-    "terser": "^5.2.1",
+    "rollup": "^2.28.2",
     "typescript": "^4.0.2"
   }
 }
